+++ backend/src/main/java/com/woori/wonfit/member/member/controller/MemberController.java
@@ -5,8 +5,11 @@ import com.woori.wonfit.member.member.dto.*;
 import com.woori.wonfit.member.member.service.MemberServiceImpl;
 import com.woori.wonfit.member.memberinfo.service.MemberInfoService;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 
 import javax.servlet.http.HttpServletRequest;
@@ -15,6 +18,7 @@ import java.util.List;
 @RestController
 @RequestMapping("/member")
 @RequiredArgsConstructor
+@Slf4j
 public class MemberController {
     private final MemberServiceImpl memberService;
     private final MemberInfoService memberInfoService;
@@ -46,8 +50,9 @@ public class MemberController {
     }
 
     @PostMapping("/leave")
-    public ResponseEntity<String> leaveMember(@RequestBody MemberLoginRequest memberLoginRequest){
+    public ResponseEntity<String> leaveMember(@RequestBody MemberLoginRequest memberLoginRequest, Authentication authentication){
         String message = memberService.leaveMember(memberLoginRequest.getLoginId(),memberLoginRequest.getPassword());
+
 
         return new ResponseEntity<>(message, HttpStatus.OK);
