--- backend/src/main/java/com/woori/wonfit/member/member/domain/JwtUtil.java
@@ -1,25 +1,0 @@
-package com.woori.wonfit.member.member.domain;
-
-import io.jsonwebtoken.Claims;
-import io.jsonwebtoken.Jwts;
-import io.jsonwebtoken.SignatureAlgorithm;
-import io.jsonwebtoken.security.Keys;
-
-import java.security.Key;
-import java.util.Date;
-
-public class JwtUtil {
-    private static final Key SECRET_KEY = Keys.secretKeyFor(SignatureAlgorithm.HS256);
-
-    public static String createToken(String memberLoginId, long expireTimeMs) {
-        Claims claims = Jwts.claims();
-        claims.put("memberLoginId", memberLoginId);
-
-        return Jwts.builder()
-                .setClaims(claims)
-                .setIssuedAt(new Date(System.currentTimeMillis()))
-                .setExpiration(new Date(System.currentTimeMillis() + expireTimeMs))
-                .signWith(SECRET_KEY)
-                .compact();
-    }
-}
